+++ AmigaOS_NDK_3.1/Includes_Libs/include_h/clib/battclock_protos.h
@@ -16,5 +16,5 @@
 #endif
 void ResetBattClock( void );
 ULONG ReadBattClock( void );
-void WriteBattClock( unsigned long time );
+void WriteBattClock( uint32_t time );
 #endif	 /* CLIB_BATTCLOCK_PROTOS_H */
