+++ AmigaOS_NDK_3.1/Includes_Libs/include_h/clib/exec_protos.h
@@ -18,41 +18,41 @@
 #include <exec/tasks.h>
 #endif
 /*------ misc ---------------------------------------------------------*/
-ULONG Supervisor( unsigned long (*userFunction)() );
+ULONG Supervisor( uint32_t (*userFunction)() );
 /*------ special patchable hooks to internal exec activity ------------*/
 /*------ module creation ----------------------------------------------*/
-void InitCode( unsigned long startClass, unsigned long version );
-void InitStruct( APTR initTable, APTR memory, unsigned long size );
+void InitCode( uint32_t startClass, uint32_t version );
+void InitStruct( APTR initTable, APTR memory, uint32_t size );
 struct Library *MakeLibrary( APTR funcInit, APTR structInit,
-	unsigned long (*libInit)(), unsigned long dataSize,
-	unsigned long segList );
+	uint32_t (*libInit)(), uint32_t dataSize,
+	uint32_t segList );
 void MakeFunctions( APTR target, APTR functionArray,
-	unsigned long funcDispBase );
+	uint32_t funcDispBase );
 struct Resident *FindResident( UBYTE *name );
-APTR InitResident( struct Resident *resident, unsigned long segList );
+APTR InitResident( struct Resident *resident, uint32_t segList );
 /*------ diagnostics --------------------------------------------------*/
-void Alert( unsigned long alertNum );
-void Debug( unsigned long flags );
+void Alert( uint32_t alertNum );
+void Debug( uint32_t flags );
 /*------ interrupts ---------------------------------------------------*/
 void Disable( void );
 void Enable( void );
 void Forbid( void );
 void Permit( void );
-ULONG SetSR( unsigned long newSR, unsigned long mask );
+ULONG SetSR( uint32_t newSR, uint32_t mask );
 APTR SuperState( void );
 void UserState( APTR sysStack );
-struct Interrupt *SetIntVector( long intNumber, struct Interrupt *interrupt );
-void AddIntServer( long intNumber, struct Interrupt *interrupt );
-void RemIntServer( long intNumber, struct Interrupt *interrupt );
+struct Interrupt *SetIntVector( int32_t intNumber, struct Interrupt *interrupt );
+void AddIntServer( int32_t intNumber, struct Interrupt *interrupt );
+void RemIntServer( int32_t intNumber, struct Interrupt *interrupt );
 void Cause( struct Interrupt *interrupt );
 /*------ memory allocation --------------------------------------------*/
-APTR Allocate( struct MemHeader *freeList, unsigned long byteSize );
+APTR Allocate( struct MemHeader *freeList, uint32_t byteSize );
 void Deallocate( struct MemHeader *freeList, APTR memoryBlock,
-	unsigned long byteSize );
-APTR AllocMem( unsigned long byteSize, unsigned long requirements );
-APTR AllocAbs( unsigned long byteSize, APTR location );
-void FreeMem( APTR memoryBlock, unsigned long byteSize );
-ULONG AvailMem( unsigned long requirements );
+	uint32_t byteSize );
+APTR AllocMem( uint32_t byteSize, uint32_t requirements );
+APTR AllocAbs( uint32_t byteSize, APTR location );
+void FreeMem( APTR memoryBlock, uint32_t byteSize );
+ULONG AvailMem( uint32_t requirements );
 struct MemList *AllocEntry( struct MemList *entry );
 void FreeEntry( struct MemList *entry );
 /*------ lists --------------------------------------------------------*/
@@ -68,15 +68,15 @@ struct Node *FindName( struct List *list, UBYTE *name );
 APTR AddTask( struct Task *task, APTR initPC, APTR finalPC );
 void RemTask( struct Task *task );
 struct Task *FindTask( UBYTE *name );
-BYTE SetTaskPri( struct Task *task, long priority );
-ULONG SetSignal( unsigned long newSignals, unsigned long signalSet );
-ULONG SetExcept( unsigned long newSignals, unsigned long signalSet );
-ULONG Wait( unsigned long signalSet );
-void Signal( struct Task *task, unsigned long signalSet );
-BYTE AllocSignal( long signalNum );
-void FreeSignal( long signalNum );
-LONG AllocTrap( long trapNum );
-void FreeTrap( long trapNum );
+BYTE SetTaskPri( struct Task *task, int32_t priority );
+ULONG SetSignal( uint32_t newSignals, uint32_t signalSet );
+ULONG SetExcept( uint32_t newSignals, uint32_t signalSet );
+ULONG Wait( uint32_t signalSet );
+void Signal( struct Task *task, uint32_t signalSet );
+BYTE AllocSignal( int32_t signalNum );
+void FreeSignal( int32_t signalNum );
+LONG AllocTrap( int32_t trapNum );
+void FreeTrap( int32_t trapNum );
 /*------ messages -----------------------------------------------------*/
 void AddPort( struct MsgPort *port );
 void RemPort( struct MsgPort *port );
@@ -90,14 +90,14 @@ void AddLibrary( struct Library *library );
 void RemLibrary( struct Library *library );
 struct Library *OldOpenLibrary( UBYTE *libName );
 void CloseLibrary( struct Library *library );
-APTR SetFunction( struct Library *library, long funcOffset,
-	unsigned long (*newFunction)() );
+APTR SetFunction( struct Library *library, int32_t funcOffset,
+	uint32_t (*newFunction)() );
 void SumLibrary( struct Library *library );
 /*------ devices ------------------------------------------------------*/
 void AddDevice( struct Device *device );
 void RemDevice( struct Device *device );
-BYTE OpenDevice( UBYTE *devName, unsigned long unit,
-	struct IORequest *ioRequest, unsigned long flags );
+BYTE OpenDevice( UBYTE *devName, uint32_t unit,
+	struct IORequest *ioRequest, uint32_t flags );
 void CloseDevice( struct IORequest *ioRequest );
 BYTE DoIO( struct IORequest *ioRequest );
 void SendIO( struct IORequest *ioRequest );
@@ -118,7 +118,7 @@ ULONG Procure( struct SignalSemaphore *sigSem,
 	struct SemaphoreMessage *bidMsg );
 void Vacate( struct SignalSemaphore *sigSem,
 	struct SemaphoreMessage *bidMsg );
-struct Library *OpenLibrary( UBYTE *libName, unsigned long version );
+struct Library *OpenLibrary( UBYTE *libName, uint32_t version );
 /*--- functions in V33 or higher (Release 1.2) ---*/
 /*------ signal semaphores (note funny registers)----------------------*/
 void InitSemaphore( struct SignalSemaphore *sigSem );
@@ -133,30 +133,30 @@ void RemSemaphore( struct SignalSemaphore *sigSem );
 /*------ kickmem support ----------------------------------------------*/
 ULONG SumKickData( void );
 /*------ more memory support ------------------------------------------*/
-void AddMemList( unsigned long size, unsigned long attributes, long pri,
+void AddMemList( uint32_t size, uint32_t attributes, int32_t pri,
 	APTR base, UBYTE *name );
-void CopyMem( APTR source, APTR dest, unsigned long size );
-void CopyMemQuick( APTR source, APTR dest, unsigned long size );
+void CopyMem( APTR source, APTR dest, uint32_t size );
+void CopyMemQuick( APTR source, APTR dest, uint32_t size );
 /*------ cache --------------------------------------------------------*/
 /*--- functions in V36 or higher (Release 2.0) ---*/
 void CacheClearU( void );
-void CacheClearE( APTR address, unsigned long length, unsigned long caches );
-ULONG CacheControl( unsigned long cacheBits, unsigned long cacheMask );
+void CacheClearE( APTR address, uint32_t length, uint32_t caches );
+ULONG CacheControl( uint32_t cacheBits, uint32_t cacheMask );
 /*------ misc ---------------------------------------------------------*/
-APTR CreateIORequest( struct MsgPort *port, unsigned long size );
+APTR CreateIORequest( struct MsgPort *port, uint32_t size );
 void DeleteIORequest( APTR iorequest );
 struct MsgPort *CreateMsgPort( void );
 void DeleteMsgPort( struct MsgPort *port );
 void ObtainSemaphoreShared( struct SignalSemaphore *sigSem );
 /*------ even more memory support -------------------------------------*/
-APTR AllocVec( unsigned long byteSize, unsigned long requirements );
+APTR AllocVec( uint32_t byteSize, uint32_t requirements );
 void FreeVec( APTR memoryBlock );
 /*------ V39 Pool LVOs...*/
-APTR CreatePool( unsigned long requirements, unsigned long puddleSize,
-	unsigned long threshSize );
+APTR CreatePool( uint32_t requirements, uint32_t puddleSize,
+	uint32_t threshSize );
 void DeletePool( APTR poolHeader );
-APTR AllocPooled( APTR poolHeader, unsigned long memSize );
-void FreePooled( APTR poolHeader, APTR memory, unsigned long memSize );
+APTR AllocPooled( APTR poolHeader, uint32_t memSize );
+void FreePooled( APTR poolHeader, APTR memory, uint32_t memSize );
 /*------ misc ---------------------------------------------------------*/
 ULONG AttemptSemaphoreShared( struct SignalSemaphore *sigSem );
 void ColdReboot( void );
@@ -167,8 +167,8 @@ void ChildOrphan( APTR tid );
 void ChildStatus( APTR tid );
 void ChildWait( APTR tid );
 /*------ future expansion ---------------------------------------------*/
-APTR CachePreDMA( APTR address, ULONG *length, unsigned long flags );
-void CachePostDMA( APTR address, ULONG *length, unsigned long flags );
+APTR CachePreDMA( APTR address, ULONG *length, uint32_t flags );
+void CachePostDMA( APTR address, ULONG *length, uint32_t flags );
 /*------ New, for V39*/
 /*--- functions in V39 or higher (Release 3) ---*/
 /*------ Low memory handler functions*/
